# resize pixelTriplet kernels