@tresreid 2022-05-11
validation including PR #143:  https://www.classe.cornell.edu/~mgr85/www/SegmentLinking_keep/PR152/eff_plots__GPU_explicit_cache_516229d_PU200/